agent/requirements.txt
langchain
langchain-openai
langgraph
langsmith
python-dotenv
requests
